requirements.txt
ffmpeg-python==0.2.0
openai==0.27.4
tiktoken==0.3.1
# PyQt5==5.15.7
python_dotenv==1.0.0
openai-whisper==20230314
